pyproject.toml
[project]
name = "nonebot-plugin-nailongmagic"
dynamic = ["version"]
description = "通过指令将目标图片变成奶龙"
authors = [
    { name = "445", email = "2877834692@qq.com" },
]
requires-python = "<4.0,>=3.9"
dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-plugin-alconna>=0.53.1",
    "nonebot-plugin-uninfo>=0.5.0",
    "pillow>=9",
    "cookit[pydantic]>=0.8.1",
    "huggingface-hub>=0.26.2",
    "torch>=2.4",
    "torchvision>=0.19",
    "diffusers>=0.31.0",
    "transformers>=4.47.0",
    "accelerate>=1.2.0",
    "peft>=0.14.0",
]
license = { text = "MIT" }
readme = "README.md"
keywords = ["nonebot", "plugin", "image"]

[project.urls]
homepage = "https://github.com/Refound-445/nonebot-plugin-nailongmagic"
repository = "https://github.com/Refound-445/nonebot-plugin-nailongmagic"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_nailongmagic/__init__.py"
